requirements.txt
absl-py==1.4.0
appdirs==1.4.4
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cloudpickle==2.2.1
docker-pycreds==0.4.0
Farama-Notifications==0.0.4
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
gymnasium==0.29.0
idna==3.4
importlib-metadata==6.8.0
Markdown==3.4.3
MarkupSafe==2.1.3
numpy==1.24.4
oauthlib==3.2.2
pandas==2.0.3
pathtools==0.1.2
protobuf==4.23.4
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sentry-sdk==1.28.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wandb==0.15.6
Werkzeug==2.3.6
zipp==3.16.2
